Updated main to call the plotting function
I'm unsure if plot_ stocks should be called in the loop or not so I opted for not. it's still in beta  there probably needs to be tested some more. I'd like to see automatic updating so we aren't just waiting for the loop to finish. i'd also like for this to be a new window but im not there yet.